get rid of warnings

+++ src/ML/SGSmooth.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 // savitzky golay smoothing.
-std::vector<double> sg_smooth(const std::vector<double>& v, const int w, const int deg);
+std::vector<double> sg_smooth(const std::vector<double>& v, const size_t w, const size_t deg);
 //! numerical derivative based on savitzky golay smoothing.
 std::vector<double> sg_derivative(const std::vector<double>& v, const int w,
     const int deg, const double h = 1.0);
